Code Import
I've sorted all my matlab files from the summer so far into something of a useful directory, and updated all? of the functions to work withing the 'METALS-CU/' directory. Not all the scripts are updated, but 'sample18_analysis.m' works and so could be useful example code to get other scripts working. It's mostly just about adding correct paths.